+++ COT/tests/test_add_disk.py
@@ -747,6 +747,8 @@ vmdk.html#streamOptimized" />
         self.assertLogged(**self.OVERWRITING_FILE)
         self.assertLogged(**self.OVERWRITING_DISK)
         self.instance.finished()
+        self.assertLogged(**self.invalid_hardware_warning(
+            "howlongofaprofilenamecanweusehere", "0", "RAM"))
         self.assertLogged(msg="Removing unused network")
         self.check_diff(file1=self.invalid_ovf, expected="""
    <ovf:References>